backend/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
python-multipart
psycopg2-binary